+++ src/main/java/at/hannibal2/skyhanni/features/chat/PlayerMessageChannel.kt
@@ -1,12 +1,17 @@
 package at.hannibal2.skyhanni.features.chat
 
-enum class PlayerMessageChannel(val prefix: String, val originalPrefix: String) {
+enum class PlayerMessageChannel(
+    val prefixColor: String,
+    val prefixSmall: String,
+    val prefixLarge: String,
+    val originalPrefix: String,
+) {
 
-    ALL("§fA>", ""),
-    ALL_GUESTING("§aA>", "§a[✌] "),
-    ALL_DUNGEON_DEAD("§7A>", "§7[GHOST] "),
-    PARTY("§9P>", "§9Party §8> "),
-    GUILD("§2G>", "§2Guild > "),
-    COOP("§bCC>", "§bCo-op > "),
+    ALL("§f", "A", "", ""),
+    ALL_GUESTING("§a", "g", "Guest", "§a[✌] "),
+    ALL_DUNGEON_DEAD("§7", "D", "Dead", "§7[GHOST] "),
+    PARTY("§9", "P", "Party", "§9Party §8> "),
+    GUILD("§2", "G", "Guild", "§2Guild > "),
+    COOP("§b", "CC", "Co-op", "§bCo-op > "),
 
 }